Updates Audioplayer Cronjob + Licensevalidation

+++ opt/max2play/update_max2play_addscripts.sh
@@ -159,13 +159,21 @@ if [ "$HW_RASPBERRY" -gt "0" ]; then
 		sudo sed -i "s/fstype=vfat,gid=users,uid=pi/fstype=vfat,gid=users,uid=pi,iocharset=iso8859-1/" /etc/usbmount/usbmount.conf
 	fi
 	
-	# Jessie Fix USBMOUNT NTFS
-	if [ -e /etc/systemd/system -a ! -e /etc/systemd/system/usbmount@.service ]; then
+	# Jessie Fix USBMOUNT NTFS ONLY ON JESSIE!
+	ISJESSIE=$(lsb_release -r | grep '8.0' | wc -l)	
+	if [ "$ISJESSIE" -gt "0" -a -e /etc/systemd/system -a ! -e /etc/systemd/system/usbmount@.service ]; then
 		echo "Fix USB-Mount on Debian Jessie"
 		echo "[Unit]\nBindTo=%i.device\nAfter=%i.device\n\n[Service]\nType=oneshot\nTimeoutStartSec=0\nEnvironment=DEVNAME=%I\nExecStart=/usr/share/usbmount/usbmount add\nRemainAfterExit=yes" > /etc/systemd/system/usbmount@.service
 		echo "# Rules for USBmount -*- conf -*-\nKERNEL==\"sd*\", DRIVERS==\"sbp2\",         ACTION==\"add\",  PROGRAM=\"/bin/systemd-escape -p --template=usbmount@.service \$env{DEVNAME}\", ENV{SYSTEMD_WANTS}+=\"%c\"\nKERNEL==\"sd*\", SUBSYSTEMS==\"usb\",       ACTION==\"add\",  PROGRAM=\"/bin/systemd-escape -p --template=usbmount@.service \$env{DEVNAME}\", ENV{SYSTEMD_WANTS}+=\"%c\"\nKERNEL==\"ub*\", SUBSYSTEMS==\"usb\",       ACTION==\"add\",  PROGRAM=\"/bin/systemd-escape -p --template=usbmount@.service \$env{DEVNAME}\", ENV{SYSTEMD_WANTS}+=\"%c\"\nKERNEL==\"sd*\",                          ACTION==\"remove\",       RUN+=\"/usr/share/usbmount/usbmount remove\"\nKERNEL==\"ub*\",                          ACTION==\"remove\",       RUN+=\"/usr/share/usbmount/usbmount remove\"" > /etc/udev/rules.d/usbmount.rules
 		rm /lib/udev/rules.d/usbmount.rules
-	fi 
+	fi
+	# Fix for NOT JESSIE and deleted usbmount rules
+	if [ "$ISJESSIE" -lt "1" -a ! -e /lib/udev/rules.d/usbmount.rules ]; then
+		echo "Remove Fix for USB-Mount on NON-Jessie"
+		echo "KERNEL==\"sd*\", DRIVERS==\"sbp2\",		ACTION==\"add\",	RUN+=\"/usr/share/usbmount/usbmount add\"\nKERNEL==\"sd*\", SUBSYSTEMS==\"usb\",	ACTION==\"add\",	RUN+=\"/usr/share/usbmount/usbmount add\"\nKERNEL==\"ub*\", SUBSYSTEMS==\"usb\",	ACTION==\"add\",	RUN+=\"/usr/share/usbmount/usbmount add\"\nKERNEL==\"sd*\",				ACTION==\"remove\",	RUN+=\"/usr/share/usbmount/usbmount remove\"\nKERNEL==\"ub*\",				ACTION==\"remove\",	RUN+=\"/usr/share/usbmount/usbmount remove\"" > /lib/udev/rules.d/usbmount.rules
+		rm /etc/systemd/system/usbmount@.service
+		rm /etc/udev/rules.d/usbmount.rules
+	fi
 fi
 
 #htaccess Password Protection Overwrite Backup
